Make: set default GHC command to use cabal environment
By running `cabal exec -- ghc` instead of standalone `ghc`, the cabal
sandbox will automatically be used if it exists.
If a sandbox doesn't exist, it should work exactly as before.

I tested this with and without a sandbox and it worked for me, but let
me know if it causes problems -- or just revert it.