Quoted identifiers: special syntax to parse 'invalid' identifiers
Because Salt programs are generally going to be machine-generated, it
doesn't matter much whether the identifier names allow underscores or
allow hyphens. The generated AST is unambiguous.
We still want an external format which is both human and machine
readable though.

I have introduced the syntax
> ##Var"quoted string"
for identifiers that can contain anything: spaces, escaped quotes, etc.

We also have:
> ##Prm"add", ##Con"Nil" and ##Sym"Symbol"