Prim: moving type params from tsig to new tpms field.
This is the first step in a series to allow Prim `teff` to share the
same typing environment as `tsig`.

This is important in order to support effects on regions, such as the
memory capabilities `Read r`, `Write r`, and `Alloc r`.

Consider the type for a primitive to read from a pointer:
    readPtr : ∀r:Region, a:Data. Ptr r a -> a ! Read r
The `r` in the effect `Read r` should be the same as in the data type
`Ptr r a -> a `.